Increased precision incase iOS moves from float to double for CGFloat values

+++ RSColorPicker/ColorPickerClasses/RSColorPickerView.m
@@ -12,29 +12,29 @@
 // Concept-code from http://www.easyrgb.com/index.php?X=MATH&H=21#text21
 BMPixel pixelFromHSV(CGFloat H, CGFloat S, CGFloat V) {
     if (S == 0) {
-        return BMPixelMake(V, V, V, 1.0f);
+        return BMPixelMake(V, V, V, 1.0);
     }
-    CGFloat var_h = H * 6.0f;
-    if (var_h == 6) {
-        var_h = 0.0f;
+    CGFloat var_h = H * 6.0;
+    if (var_h == 6.0) {
+        var_h = 0.0;
     }
-    CGFloat var_i = floorf(var_h);
-    CGFloat var_1 = V * (1 - S);
-    CGFloat var_2 = V * (1 - S * (var_h - var_i));
-    CGFloat var_3 = V * (1 - S * (1 - (var_h - var_i)));
+    CGFloat var_i = floor(var_h);
+    CGFloat var_1 = V * (1.0 - S);
+    CGFloat var_2 = V * (1.0 - S * (var_h - var_i));
+    CGFloat var_3 = V * (1.0 - S * (1.0 - (var_h - var_i)));
     
     if (var_i == 0) {
-        return BMPixelMake(V, var_3, var_1, 1.0f);
+        return BMPixelMake(V, var_3, var_1, 1.0);
     } else if (var_i == 1) {
-        return BMPixelMake(var_2, V, var_1, 1.0f);
+        return BMPixelMake(var_2, V, var_1, 1.0);
     } else if (var_i == 2) {
-        return BMPixelMake(var_1, V, var_3, 1.0f);
+        return BMPixelMake(var_1, V, var_3, 1.0);
     } else if (var_i == 3) {
-        return BMPixelMake(var_1, var_2, V, 1.0f);
+        return BMPixelMake(var_1, var_2, V, 1.0);
     } else if (var_i == 4) {
-        return BMPixelMake(var_3, var_1, V, 1.0f);
+        return BMPixelMake(var_3, var_1, V, 1.0);
     }
-    return BMPixelMake(V, var_1, var_2, 1.0f);
+    return BMPixelMake(V, var_1, var_2, 1.0);
 }
 
 
@@ -49,7 +49,7 @@ BMPixel pixelFromHSV(CGFloat H, CGFloat S, CGFloat V) {
 
 - (id)initWithFrame:(CGRect)frame
 {
-    CGFloat sqr = fminf(frame.size.height, frame.size.width);
+    CGFloat sqr = fmin(frame.size.height, frame.size.width);
     frame.size = CGSizeMake(sqr, sqr);
     
     self = [super initWithFrame:frame];
@@ -57,15 +57,15 @@ BMPixel pixelFromHSV(CGFloat H, CGFloat S, CGFloat V) {
         cropToCircle = YES;
         
         selection = CGPointMake(sqr/2, sqr/2);
-        selectionView = [[UIView alloc] initWithFrame:CGRectMake(0.0f, 0.0f, 18.0f, 18.0f)];
+        selectionView = [[UIView alloc] initWithFrame:CGRectMake(0.0, 0.0, 18.0, 18.0)];
         selectionView.backgroundColor = [UIColor clearColor];
-        selectionView.layer.borderWidth = 2.0f;
-        selectionView.layer.borderColor = [UIColor colorWithWhite:0.1f alpha:1.0f].CGColor;
-        selectionView.layer.cornerRadius = 9.0f;
+        selectionView.layer.borderWidth = 2.0;
+        selectionView.layer.borderColor = [UIColor colorWithWhite:0.1 alpha:1.0].CGColor;
+        selectionView.layer.cornerRadius = 9.0;
         [self updateSelectionLocation];
         [self addSubview:selectionView];
         
-        self.brightness = 1.0f;
+        self.brightness = 1.0;
         rep = [[ANImageBitmapRep alloc] initWithSize:BMPointFromSize(frame.size)];
     }
     return self;
@@ -84,9 +84,9 @@ BMPixel pixelFromHSV(CGFloat H, CGFloat S, CGFloat V) {
 }
 
 -(void)genBitmap {
-    CGFloat radius = (self.frame.size.width / 2);
-    CGFloat relX = 0;
-    CGFloat relY = 0;
+    CGFloat radius = (self.frame.size.width / 2.0);
+    CGFloat relX = 0.0;
+    CGFloat relY = 0.0;
     
     for (int x = 0; x < self.frame.size.width; x++) {
         relX = x - radius;
@@ -94,17 +94,17 @@ BMPixel pixelFromHSV(CGFloat H, CGFloat S, CGFloat V) {
         for (int y = 0; y < self.frame.size.height; y++) {
             relY = radius - y;
             
-            CGFloat r_distance = sqrtf((relX * relX)+(relY * relY));
+            CGFloat r_distance = sqrt((relX * relX)+(relY * relY));
             if (fabsf(r_distance) > radius && cropToCircle == YES) {
-                [rep setPixel:BMPixelMake(0.0f, 0.0f, 0.0f, 0.0f) atPoint:BMPointMake(x, y)];
+                [rep setPixel:BMPixelMake(0.0, 0.0, 0.0, 0.0) atPoint:BMPointMake(x, y)];
                 continue;
             }
-            r_distance = fminf(r_distance, radius);
+            r_distance = fmin(r_distance, radius);
             
-            CGFloat angle = atan2f(relY, relX);
-            if (angle < 0) { angle = (2*M_PI)+angle; }
+            CGFloat angle = atan2(relY, relX);
+            if (angle < 0.0) { angle = (2.0 * M_PI)+angle; }
             
-            CGFloat perc_angle = (angle/(2 * M_PI));
+            CGFloat perc_angle = angle / (2.0 * M_PI);
             BMPixel thisPixel = pixelFromHSV(perc_angle, r_distance/radius, self.brightness);
             [rep setPixel:thisPixel atPoint:BMPointMake(x, y)];
         }
@@ -137,7 +137,7 @@ BMPixel pixelFromHSV(CGFloat H, CGFloat S, CGFloat V) {
     if (![self pointInside:point withEvent:event]) { return; }
     
     BMPixel pixel = [rep getPixelAtPoint:BMPointFromPoint(point)];
-    if (pixel.alpha > 0.0f) {
+    if (pixel.alpha > 0.0) {
         badTouch = NO;
         selection = point;
         [delegate colorPickerDidChangeSelection:self];
@@ -152,7 +152,7 @@ BMPixel pixelFromHSV(CGFloat H, CGFloat S, CGFloat V) {
     CGPoint point = [[touches anyObject] locationInView:self];
     if (![self pointInside:point withEvent:event]) { return; }
     BMPixel pixel = [rep getPixelAtPoint:BMPointFromPoint(point)];
-    if (pixel.alpha > 0.0f) {
+    if (pixel.alpha > 0.0) {
         selection = point;
         [delegate colorPickerDidChangeSelection:self];
         [self updateSelectionLocation];
@@ -163,7 +163,7 @@ BMPixel pixelFromHSV(CGFloat H, CGFloat S, CGFloat V) {
     CGPoint point = [[touches anyObject] locationInView:self];
     if (![self pointInside:point withEvent:event]) { return; }
     BMPixel pixel = [rep getPixelAtPoint:BMPointFromPoint(point)];
-    if (pixel.alpha > 0.0f) {
+    if (pixel.alpha > 0.0) {
         selection = point;
         [delegate colorPickerDidChangeSelection:self];
         [self updateSelectionLocation];
